Initial commit: In the beginning was the Word, and the Word was 'cargo new'
Like the Librarian organizing books that haven't been written yet, this
repository springs forth fully-formed with the inexorable logic of
someone who has read the manual but hasn't quite grasped that the
manual was written by committee.

Contains:
- A Rust crate that speaks to Google's Gemini with the confidence of
  someone asking for directions in a foreign language
- Nix flakes, because apparently we needed MORE ways to make things
  reproducible (the Patrician would approve)
- A gitignore file that excludes everything important and includes
  everything that shouldn't be there, in the finest tradition of
  institutional knowledge

"It's not a bug, it's an undocumented feature," as they say in
Ankh-Morpork's Guild of Software Engineers (Motto: "We Get There
Eventually").